==> 3.transformers_similarities.ipynb <==
## Semantic Similarity

from google.colab import drive
drive.mount('/content/drive')

!pip install sentence_transformers

import torch 
print(torch.cuda.device_count())

torch.cuda.get_device_name(0)

import pandas as pd 

df = pd.read_csv('/content/drive/My Drive/arXiv/selected_papers.csv')

df

test_corpus = df[:1000].abstract.str.strip().to_list()

df.loc[270]

len(paraphrases)

len(paraphrases)

len(test_corpus) * len(test_corpus)

from sentence_transformers import SentenceTransformer, util

model = SentenceTransformer('distilbert-base-nli-stsb-mean-tokens')

paraphrases = util.paraphrase_mining(model, test_corpus, top_k=20)

#for paraphrase in paraphrases:
 #   score, i, j = paraphrase
   # print("{} \t\t {} \t\t Score: {:.4f}".format(test_corpus[i], test_corpus[j], score))
    #print("{} \t\t {} \t\t Score: {:.4f}".format(i, j, score))


whole_corpus = df.abstract.str.strip().to_list()

from sentence_transformers import SentenceTransformer, util

model = SentenceTransformer('distilbert-base-nli-stsb-mean-tokens')

paraphrases_whole = util.paraphrase_mining(model, whole_corpus,len(whole_corpus), top_k=100)

#for paraphrase in paraphrases:
 #   score, i, j = paraphrase
   # print("{} \t\t {} \t\t Score: {:.4f}".format(test_corpus[i], test_corpus[j], score))
    #print("{} \t\t {} \t\t Score: {:.4f}".format(i, j, score))


len(paraphrases_whole) 

len(paraphrases_whole)

paraphrases_whole[-1]

paraphrases_whole[944]

len(paraphrases_whole)

df.loc[463], df.loc[499]

paraphrases[-1]

similarities_df = pd.DataFrame(paraphrases_whole, columns = ['similarities', 'paper1_idx', 'paper2_idx'])  

tmp = similarities_df.merge(df[['id','ref']], how='left', left_on='paper1_idx', right_on='ref').drop('ref', axis =1).rename(columns={'id': 'Source', 'similarities':'Weight'})

edges = tmp.merge(df[['id','ref']], how='left', left_on='paper2_idx',right_on='ref').drop(['ref','paper1_idx','paper2_idx'], axis =1).rename(columns={'id': 'Target'})

edges

#nodes.to_csv('/content/drive/My Drive/arXiv/transformers_nodes.csv', index=False)
#edges.to_csv('/content/drive/My Drive/arXiv/transformers_edges.csv', index=False)

df2 = pd.read_csv('/content/drive/My Drive/arXiv/transformers_edges.csv')

reference = list(set(df2.Source.astype(str).to_list() + df2.Target.astype(str).to_list()) ) 

nodes = df[df['id'].isin(reference)].rename(columns={'id': 'Id'}).drop('abstract', axis=1)

nodes['Label'] = nodes.Id
nodes.to_csv('/content/drive/My Drive/arXiv/transformers_nodes.csv', index=False)

You will get for each sentence only the one most other relevant sentence. Note, if B is the most similar sentence for A, A must not be the most similar sentence for B. So it can happen that the returned list contains entries like (A, B) and (B, C).
