==> 5.citations_formatter.ipynb <==
from pathlib import Path
import pandas as pd


#batches were ran in colab as well as local machine for time constraints
# as the api limit was only ~150 for 5 minutes. 
# However, there might be overlap of results so it needs to be filtered/processed

data_dir1 = Path('citations_batchrun_1/')
colab_dfs = pd.concat(
    pd.read_csv(csv_file,dtype=object)
    for csv_file in data_dir1.glob('*.csv')
)


data_dir2 = Path('citations_batchrun2/')
local_dfs = pd.concat(
    pd.read_csv(csv_file,dtype=object)
    for csv_file in data_dir2.glob('*.csv')
)


#full_df.to_csv('similarities.csv', index=True)

df = local_dfs.append(colab_dfs, ignore_index=True)

len(df)

df.columns = ['Target', 'Source']

len(df)

edges = df.drop_duplicates(['Target', 'Source'])

len(edges)

selected_papers = pd.read_csv('selected_papers.csv')

len(selected_papers)

reference = list(set(edges.Source.astype(str).to_list() + edges.Target.astype(str).to_list()) ) 

len(reference)

nodes = selected_papers[selected_papers['id'].isin(reference)].drop('abstract', axis=1).rename(columns={'id': 'Id', 'title':'Title'})
nodes['Label'] = nodes['Id']

nodes.dtypes

nodes = selected_papers[selected_papers['id'].isin(reference)].drop('abstract', axis=1).rename(columns={'id': 'Id', 'title':'Title'})
nodes['Label'] = nodes['Id']

edges.to_csv('gephi/citations_network_edges.csv', index=False)


len(edges)

len(nodes)

extracted_topics = pd.read_csv('gephi/extracted_topics.csv')

extracted_topics.head(10)

len(extracted_topics)

len(nodes)

nodes = nodes.merge(extracted_topics, left_on='Id', right_on='id').drop(['title','id'], axis=1)

nodes.to_csv('gephi/citations_network_nodes.csv', index=False)
