==> 7.data_for_Neo4j.ipynb <==
from google.colab import drive
drive.mount("/content/drive", force_remount=True)

!python -m pip install dask
!python -m pip install 'fsspec>=0.3.3'

import numpy as np 
import pandas as pd
import dask.bag as db
import json

lines=db.read_text('/content/drive/My Drive/arXiv/metadata.json')
lines 

records=lines.map(lambda x:json.loads(x))
records.take(4)

records_count=records.count()
print("Number of Records in ArXiv Data is ",records_count.compute())

ai_category_list=['cs.LG','cs.AI','cs.CV']
ai_docs = (records.filter(lambda x:any(ele in x['categories'] for ele in ai_category_list)==True))
print("Total Papers published in AI&ML ",ai_docs.count().compute())

ai_docs.take(1)

extract_latest_version_year=lambda x:x['versions'][-1]["created"].split(" ")[3]

ai_docs_by_year=ai_docs.map(extract_latest_version_year).frequencies().to_dataframe(columns=['submission_year','num_submissions']).compute()

df = df.filter(df.year>2018)

get_metadata = lambda x: {'id': x['id'].replace('\n',' '),
                  'title': x['title'].replace('\n',''),
                  'authors': x['authors'].replace('\n',' '),
                  'submitter' : x['submitter'],
                  'category':x['categories'].replace('\n',' '),
                 'version':x['versions'][-1]['created'].replace('\n',' '),
                 'date':x['versions'][-1]["created"],
                 'year':int(x['versions'][-1]["created"].split(" ")[3])}

docs = ai_docs.map(get_metadata)

data_after_2018=docs.filter(lambda x:x['year']>2018)

data_after_2018.count().compute()

data_after_2018.take(1)

df = data_after_2018.to_dataframe().compute()

df.head()

df?

df.dtypes

len(df)

 df['authors'] = df['authors'].str.replace("\\","").str.replace("and",",").str.split(",")

df['authors']=df['authors'].apply(lambda x: [y.strip() for y in x])

 df['category'] = df['category'].str.split(" ")

import datetime 

datetime.datetime.strptime('Mon Feb 15 2010', '%a %b %d %Y').strftime('%d/%m/%Y')
'15/02/2010'

df['date'] = df['date'].str[5:16]
df['date'] =  pd.to_datetime(df['date'])

df.drop('version', axis=1, inplace=True)

df.head() 

df.loc[df.id=='2005.14223']

#df.to_dict('records')

similarities_transformers = pd.read_csv('/content/drive/My Drive/arXiv/transformers_edges.csv',dtype='object')

similarities_transformers.dtypes

citations = pd.read_csv('/content/drive/My Drive/arXiv/citations_network_edges.csv', dtype='object')

#target_cited_by_source

citations.dtypes

citations.Target.value_counts()

topics = pd.read_csv('/content/drive/My Drive/arXiv/extracted_topics.csv')

topics.head()

main = df.merge(topics, on='id')

citations.columns

citations.dtypes

citations.head(10)

%%time
citations_as_list = citations.groupby('Target')['Source'].apply(list).reset_index(name='cited_by')

citations_as_list.columns

len(df) - len(citations_as_list)

pipe_df = main.merge(citations_as_list, how='left', left_on='id', right_on='Target').drop(['Target','title_y'], axis=1)

similarities_transformers.columns

%%time
similarities_as_list = similarities_transformers.groupby('Source')['Target'].apply(list).reset_index(name='similar_to')

similarities_as_list.head()

pipe_df_2 = pipe_df.merge(similarities_as_list, how='left', left_on='id', right_on='Source').drop(['Source'], axis=1)

len(df)-len(similarities_as_list)

pipe_df_2.isna().sum()

pipe_df_2.rename({'title_x':'title'}, axis=1, inplace=True)

pipe_df_2.head()

pipe_df_2['date'] = df['date'].astype(str)

main_dict = pipe_df_2.to_dict('records')

json_object = json.dumps(main_dict)   

with open("/content/drive/My Drive/arXiv/data.json", "w") as outfile:  
    json.dump(main_dict, outfile) 

print(json.dumps(main_dict[25], indent=2))


import json

with open('/content/drive/My Drive/arXiv/data.json') as f:
  data = json.load(f)


def cleanNullTerms(d):
   return {
      k:v
      for k, v in d.items()
      if v is not None
   }


for d in data:
  cleanNullTerms(d)

import numpy as np 

res = {k:v for k,v in data.items() if v is not np.nan}
